refactor: modularize burst fitting components
